requirements.txt
streamlit
pdfplumber
tqdm
faiss-cpu
sentence-transformers
openai
python-dotenv
